[sharedUtils] move createTypeWords from handlers to sharedUtils
